pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "layro"
version = "0.1.0"
description = "A robust, reusable configuration management package for Python applications"
readme = "layro/README.md"
authors = [
    {name = "Neil", email = "your.email@example.com"},
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["configuration", "cli", "yaml", "dataclass", "machine-learning"]
dependencies = [
    "pyyaml>=6.0",
    "tyro>=0.5.0",
]
requires-python = ">=3.8"

[project.urls]
"Homepage" = "https://github.com/yourusername/layro"
"Bug Tracker" = "https://github.com/yourusername/layro/issues"

[tool.hatch.build.targets.wheel]
packages = ["layro"]

[tool.pytest.ini_options]
testpaths = ["layro"]
python_files = "test_*.py"